requirements.txt
pyserial==3.5
fastapi==0.63.0
uvicorn==0.13.3
